Update 01-introduction.md
Updated GridSearch code. Additional chunk to define a function used in the KerasClassifier functions build_fn arg. Function create_model is the previous code consolidated in the function definition to build layers, and compile on each combo of grid search params.  Initially defined with 10 epochs but compute time was around 45min. Have reduced to 2 epochs to save computation time as this mainly for demonstration purposes at this point in the workshop.